Updated Notebook Functionality
Rehauled the notebook system. Now there is a universal notebook object which controls everything. Everything is as it was before, but now it uses objects instead of DrawGUI sprites, and the bug where you couldn't close the notebook using a different button is fixed.

Notebook object uses a map data structure to store all the information regarding what elements of each plant have been discovered in the notebook to make it easier to maintain the state of the notebook throughout opening and closing it.

Also changed all the sizes of most of the sprites to fit the screen better, the notebook better, and to not take up so much space.

Notebook is still very much a WIP