Notebook Functionality for 1st Page
I hate GML lmao.

All sprites and objects and everything are setup for the first two pages. Everything for the first plant is good and lightly bug tested.

You can now open and close the notebook, click on the interactable objects and have them change color and make an entry. The notebook saves states between closing and opening it, and the objects all still move in unison with the camera